Removed the connected nodes from dashboard
Removed the connected nodes list from the dashboard page because it was
causing errors that I could not figure out. This is only a temp fix (It
was only a proof of concept anyways). I also cleaned up the view for the
dashboard because it imported the same thing twice. Right now
everything works. But is is very web1.0 and looks like a professors
website. It could use some CSS.